Fix the isse with qd and qdd
Somehow, there was and issue with sign of methods to get qd and qdd.
The problem seems to be coming from the matlab code that I wrote to
test the idea and there, I wrote qi - (qf - qi) which is wrong and
shoudl be qi + (qf - qi). I am writing this long ass commit message
because I am still not sure if this is the fix or no. All I know is
that none of the previous codes should have worked...
God bless feedback!